dsba6155project/terraform/cloud_function/requirements.txt
google-cloud-storage==1.7.0
requests==2.13.0
bs4
